requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
aniso8601==7.0.0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
channels==2.4.0
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==2.5.0
Deprecated==1.2.12
Django==3.0.3
django-channels-graphql-ws==0.8.0
django-cors-headers==3.7.0
django-emoji==2.2.2
djongo==1.3.6
emoji==0.6.0
firebase-admin==5.0.1
gcloud==0.18.3
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.39.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
graphene==2.1.8
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.38.1
hiredis==2.0.0
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
install==1.3.4
jwcrypto==0.9.1
msgpack==1.0.2
multidict==5.1.0
oauth2client==4.1.3
packaging==20.9
promise==2.3
proto-plus==1.18.1
protobuf==3.17.3
psycopg2==2.9.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pymongo==3.11.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
Pyrebase4==4.5.0
python-jwt==3.3.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.6.2
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
wrapt==1.12.1
yarl==1.6.3
zope.interface==5.4.0
